refactor : todoApp 구조 변경
