+++ src/components/todoApp/validator/validator.js
