+++ src/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import TodoApp from './todoApp.js';
+import TodoApp from './components/todoApp/todoApp.js';
 
 const todoApp = new TodoApp();
 todoApp.init();
