--- src/view/userListView.js
@@ -1,11 +1,0 @@
-'use strict';
-
-import { $ } from '../utils/dom.js';
-
-class UserListView {
-  constructor() {
-    this.$usetList = $('#user-list');
-    this.$userCreateBtn = $('.user-create-button');
-    this.$userDeleteBtn = $('.user-delete-button');
-  }
-}
