--- src/view/userTitleView.js
@@ -1,7 +1,0 @@
-'use strict';
-
-import { $ } from '../utils/dom.js';
-
-class UserTitleView {
-  constructor() {}
-}
